Properly load the cached equipment of an entity. Fixes BUKKIT-5650
When a LivingEntity is loaded, its cached attributes (For example weapon damage) are loaded from its NBT-Tag. But it is not loaded where these attributes derive from. This can produce wrong cache data, when the inventory of the player is modified before it lives a singe tick (For example in the PlayerJoinEvent). This commit fixes the issue by correctly initializing the cache before the inventory can be modified by plugins.